收藏算法修改。

+++ algm-algorithms-boot/src/main/java/com/beyond/algm/algmalgorithmsboot/controller/AlgModuleListController.java
@@ -139,7 +139,7 @@ public class AlgModuleListController extends BaseController {
     @RequestMapping(value = "/module/star", method = RequestMethod.POST, produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
     public Result AddAlgorithm(String modSn) throws AlgException {
         AlgUser algUser = getUserInfo();
-        int count = algModuleListService.modStar(modSn,algUser.getUsrSn());
+        Long count = algModuleListService.modStar(modSn,algUser.getUsrSn());
         return Result.ok(count);
     }
